nelete namespaces, edit mail notification
