--- app/views/mailers/member/screening/send_followup_mailer/notify.html.haml
@@ -1,8 +1,0 @@
-%p.lead
-  = t('dashboard.users.mailers.uncompleted.title')
-%p
-  #{t('dashboard.users.mailers.hello')} #{@role.user.full_name}
-%p
-  #{t('dashboard.users.mailers.uncompleted.text_one')}
-  = link_to t('dashboard.users.mailers.uncompleted.link'), root_url
-  #{t('dashboard.users.mailers.uncompleted.text_two')}
